move run_probe stuff into reportgen
